requirements.txt
adal==1.2.5
APScheduler==3.6.3
azure-common==1.1.26
azure-mgmt-storage==11.1.0
azure-mgmt-subscription==0.6.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
dataclasses==0.6
fastapi==0.63.0
Flask==1.1.2
h11==0.12.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
msrestazure==0.6.4
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
uvicorn==0.13.3
Werkzeug==1.0.1
